Adding complete Espressif IDF compatibility

+++ src/Helper.cpp
@@ -2,13 +2,12 @@
 #include "Helper.h"
 
 // Local includes.
-#include "Configuration.h"
+#include "Constants.h"
 
 // Library includes.
 #include <stdarg.h>
 #include <stdio.h>
 #include <string.h>
-#include <pgmspace.h>
 #include <assert.h>
 
 uint8_t Helper::detectSize(const char *msg, ...) {
@@ -21,15 +20,14 @@ uint8_t Helper::detectSize(const char *msg, ...) {
       assert(result >= 0);
       va_end(args);
       return result;
-
 }
 
-uint32_t Helper::getOccurences(const char *str, char symbol) {
-    uint32_t count = 0;
+size_t Helper::getOccurences(const char *str, char symbol) {
+    size_t count = 0;
     if (str == nullptr) {
       return count;
     }
-    for (uint32_t i = 0; i < strlen(str); i++) {
+    for (size_t i = 0; i < strlen(str); i++) {
       if (str[i] != symbol) {
         continue;
       }
